provides full support for quoting handling in data ; this is a large bugfix
it's complex, dumb and really annoying. This whole thing is necessary…
because we have to choose between "quote everything" or "quote nothing",
with the obvious problem that not quoting biological data is dangerous.
So we have to quote everything, which is quite verbose, hence the
modification of tests to ensure quoting.
Fortunately, it's just a matter of detecting non quoted data now if ASP
is given directly.
Arg. I hate this.