example showing limits of internal SCC search
@cfrioux
Because an SCC is treated as a whole object, only perfectly optimized
solutions are provided. As a consequence, even with pareto search,
some solutions are discarded for quite obscur reasons.

Without pareto, we get only `d` as solution, because most of the other
solutions are in the same SCC, and therefore discarded as not optimal
enough.

See comments for more explanations and some solutions ideas.